Move hardware info into separate pal header
