+++ include/ccf/crypto/entropy.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "ccf/pal/mem.h"
+#include "ccf/pal/hardware_info.h"
 
 #include <cassert>
 #include <cstdint>
